Added Facebook register option for website
